Move public macros to new macros module
